+++ src/modules/tools/compareFeatures/components/ComparisonList.vue
@@ -88,6 +88,16 @@ export default {
                                     target="_blank"
                                 >{{ value }}</a>
                             </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'true'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.true') }}</span>
+                            </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'false' || value.toLowerCase() === 'no'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.false') }}</span>
+                            </p>
                             <p v-else-if="isPhoneNumber(value)">
                                 <a :href="getPhoneNumberAsWebLink(value)">{{ value }}</a>
                             </p>
@@ -140,6 +150,16 @@ export default {
                                     :href="value"
                                     target="_blank"
                                 >{{ value }}</a>
+                            </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'true'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.true') }}</span>
+                            </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'false' || value.toLowerCase() === 'no'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.false') }}</span>
                             </p>
                             <p v-else-if="isPhoneNumber(value)">
                                 <a :href="getPhoneNumberAsWebLink(value)">{{ value }}</a>
@@ -205,6 +225,16 @@ export default {
                                     target="_blank"
                                 >{{ value }}</a>
                             </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'true'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.true') }}</span>
+                            </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'false' || value.toLowerCase() === 'no'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.false') }}</span>
+                            </p>
                             <p v-else-if="isPhoneNumber(value)">
                                 <a :href="getPhoneNumberAsWebLink(value)">{{ value }}</a>
                             </p>
@@ -259,6 +289,16 @@ export default {
                                     :href="value"
                                     target="_blank"
                                 >{{ value }}</a>
+                            </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'true'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.true') }}</span>
+                            </p>
+                            <p
+                                v-else-if="value.toLowerCase() === 'false' || value.toLowerCase() === 'no'"
+                            >
+                                <span>{{ $t('common:modules.tools.compareFeatures.trueFalse.false') }}</span>
                             </p>
                             <p v-else-if="isPhoneNumber(value)">
                                 <a :href="getPhoneNumberAsWebLink(value)">{{ value }}</a>